[TASK] Split admin and model modules [BUGFIX] Fixed use of root domain for link generation
