+++ seosnap/admin/__init__.py
@@ -1,0 +1,3 @@
+from .extract_field import *
+from .page import *
+from .website import *
