+++ seosnap/models/__init__.py
@@ -1,0 +1,3 @@
+from .website import *
+from .extract_field import *
+from .page import *
