+++ seosnap/models/website.py
@@ -1,10 +1,11 @@
 from django.contrib.auth.models import AbstractUser
 from django.db import models
+from django_mysql.models import Model, QuerySet
 from django.db.models.expressions import RawSQL
-from django_mysql.models import Model
-from django_mysql.models import QuerySet
 
-from seosnap.utils.JSONField import JSONField
+from urllib.parse import urlparse
+
+import seosnap
 
 
 class Website(Model):
@@ -26,8 +27,10 @@ class Website(Model):
         return [field for field in self.extract_fields.filter(display=True).values_list('name', flat=True)]
 
     def get_url(self, path):
-        if isinstance(path, Page): path = path.address
-        return self.domain.rstrip('/') + path
+        if isinstance(path, seosnap.models.Page): path = path.address
+        uri = urlparse(self.domain)
+        root_domain = f'{uri.scheme}://{uri.netloc}'
+        return root_domain + path
 
 
 def website_add_permission_filter(qs, permission, user: AbstractUser):
@@ -45,41 +48,3 @@ def website_add_permission_filter(qs, permission, user: AbstractUser):
         WHERE ggop.permission_id IN ({select_permission}) AND ggop.group_id in ({select_groups})
     ''', (permission, user.id, permission, user.id)))
     return qs
-
-
-class ExtractField(Model):
-    website = models.ForeignKey(Website, related_name='extract_fields', on_delete=models.CASCADE)
-
-    name = models.CharField(max_length=255)
-    css_selector = models.CharField(max_length=255)
-    display = models.BooleanField(default=False)
-
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-
-    def __str__(self):
-        return f'{self.name} - {self.css_selector[:30]}'
-
-
-class Page(Model):
-    website = models.ForeignKey(Website, related_name='pages', on_delete=models.CASCADE)
-
-    address = models.CharField(max_length=255)
-    content_type = models.CharField(max_length=255, null=True, default=None)
-    status_code = models.IntegerField(null=True, default=None)
-    extract_fields = JSONField(default=dict)
-
-    cache_status = models.CharField(max_length=64, choices=[
-        ('cached', 'Cached'),
-        ('not-cached', 'Not Cached')
-    ], default='not-cached')
-    cached_at = models.DateTimeField(null=True, default=None)
-
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-
-    def __str__(self):
-        return f'{self.address}'
-
-    def get_extract_field(self, name):
-        return self.extract_fields.get(name, None)
